+++ test/regressions/test_lake_admin.py
@@ -12,63 +12,113 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-import os
 
 import pytest
-from custom_resources import CustomApiObject
+from custom_resources import OrbitJobCustomApiObject
+from common_utils import JOB_COMPLETION_STATUS
 from kubetest.client import TestClient
+import logging
 
-
-MANIFESTS_PATH = os.path.join(
-    os.path.dirname(os.path.realpath(__file__)),
-    "manifests",
+# Initialize parameters
+logging.basicConfig(
+    format="%(asctime)s %(levelname)-8s %(message)s",
+    level=logging.INFO,
+    datefmt="%Y-%m-%d %H:%M:%S",
 )
 
+logger = logging.getLogger()
 
-class LakeAdmin(CustomApiObject):
-    group = "orbit.aws"
-    api_version = "v1"
-    kind = "OrbitJob"
-
-    def is_ready(self) -> bool:
-        self.refresh()
-        return self.obj.get("status", {}).get("orbitJobOperator", {}).get("jobStatus") == "JobCreated"
-
-
-@pytest.mark.order(1)
-@pytest.mark.namespace(create=False)
-@pytest.mark.testlakeadmin
-def test_lakeadmin(kube: TestClient) -> None:
-    body = {
-        "apiVersion": "orbit.aws/v1",
-        "kind": "OrbitJob",
-        "metadata": {
-            "name": "test-orbit-job"
-        },
-        "spec": {
-            "taskType": "jupyter",
-            "compute": {
-                 "nodeType": "ec2",
-                 "container": {
-                     "concurrentProcesses": 1
-                 },
-                 "podSetting": "orbit-runner-support-large"
+LAKE_ADMIN_JOB = {
+    "apiVersion": "orbit.aws/v1",
+    "kind": "OrbitJob",
+    "metadata": {
+        "generateName": "test-orbit-job-lake-admin-"
+    },
+    "spec": {
+        "taskType": "jupyter",
+        "compute": {
+            "nodeType": "ec2",
+            "container": {
+                "concurrentProcesses": 1
             },
-            "tasks": [{
-                "notebookName": "Example-1-SQL-Analysis-Athena.ipynb",
-                "sourcePath": "shared/samples/notebooks/B-DataAnalyst",
-                "targetPath": "shared/regression/notebooks/B-DataAnalyst",
-                "targetPrefix": "yyy",
-                "params": {
-                    "glue_db": "glue_database",
-                    "target_db": "users"
-                }
-            }]
-        }
+            "podSetting": "orbit-runner-support-small"
+        },
+        "tasks": []
     }
+}
 
-    print(body)
-    lakeadmin = LakeAdmin(body)
+@pytest.mark.namespace(create=False)
+@pytest.mark.testlakeadmin_ebs
+def test_lakeadmin_1_ebs(kube: TestClient) -> None:
+
+    lake_admin_job_ebs = LAKE_ADMIN_JOB
+    lake_admin_job_ebs["metadata"]["generateName"]= "test-orbit-job-lake-admin-ebs-"
+    lake_admin_job_ebs["spec"]["tasks"]= [{
+                "notebookName": "1-EBS.ipynb",
+                "sourcePath": "shared/samples/notebooks/M-Admin",
+                "targetPath": "shared/regression/notebooks/M-Admin",
+                "params": {}
+            }]
+    logger.info(lake_admin_job_ebs)
+    lakeadmin = OrbitJobCustomApiObject(lake_admin_job_ebs)
     lakeadmin.create(namespace="lake-admin")
-    lakeadmin.wait_until_ready(timeout=30)
+    # Logic to wait till OrbitJob creates
+    lakeadmin.wait_until_ready(timeout=60)
+    # Logic to pass or fail the pytest
+    lakeadmin.wait_until_job_completes(timeout=1200)
+    current_status = lakeadmin.get_status().get("orbitJobOperator").get("jobStatus")
+    logger.info(f"current_status={current_status}")
+    #Cleanup
     lakeadmin.delete()
+    assert current_status == JOB_COMPLETION_STATUS
+
+
+@pytest.mark.namespace(create=False)
+@pytest.mark.testlakeadmin_image_with_apps
+def test_lakeadmin_2_image_with_apps(kube: TestClient) -> None:
+    lake_admin_job_mage_with_apps = LAKE_ADMIN_JOB
+    lake_admin_job_mage_with_apps["metadata"]["generateName"]= "test-orbit-job-lake-admin-image-with-apps-"
+    lake_admin_job_mage_with_apps["spec"]["tasks"]= [{
+                "notebookName": "2-Image_with_apps.ipynb",
+                "sourcePath": "shared/samples/notebooks/M-Admin",
+                "targetPath": "shared/regression/notebooks/M-Admin",
+                "params": {}
+            }]
+
+    logger.info(lake_admin_job_mage_with_apps)
+    lakeadmin = OrbitJobCustomApiObject(lake_admin_job_mage_with_apps)
+    lakeadmin.create(namespace="lake-admin")
+    # Logic to wait till OrbitJob creates
+    lakeadmin.wait_until_ready(timeout=60)
+    # Logic to pass or fail the pytest
+    lakeadmin.wait_until_job_completes(timeout=1200)
+    current_status = lakeadmin.get_status().get("orbitJobOperator").get("jobStatus")
+    logger.info(f"current_status={current_status}")
+    #Cleanup
+    lakeadmin.delete()
+    assert current_status == JOB_COMPLETION_STATUS
+
+@pytest.mark.namespace(create=False)
+@pytest.mark.testlakeadmin_lf
+def test_lakeadmin_3_lf_account_settings(kube: TestClient) -> None:
+    lake_admin_job_lf = LAKE_ADMIN_JOB
+    lake_admin_job_lf["metadata"]["generateName"]= "test-orbit-job-lake-admin-lf-account-settings-"
+    lake_admin_job_lf["spec"]["tasks"]= [{
+                "notebookName": "4-LakeFormation-Account-Settings.ipynb",
+                "sourcePath": "shared/samples/notebooks/M-Admin",
+                "targetPath": "shared/regression/notebooks/M-Admin",
+                "params": {}
+            }]
+
+    logger.info(lake_admin_job_lf)
+    lakeadmin = OrbitJobCustomApiObject(lake_admin_job_lf)
+    lakeadmin.create(namespace="lake-admin")
+    # Logic to wait till OrbitJob creates
+    lakeadmin.wait_until_ready(timeout=60)
+    # Logic to pass or fail the pytest
+    lakeadmin.wait_until_job_completes(timeout=1200)
+    current_status = lakeadmin.get_status().get("orbitJobOperator").get("jobStatus")
+    logger.info(f"current_status={current_status}")
+    #Cleanup
+    lakeadmin.delete()
+    assert current_status == JOB_COMPLETION_STATUS